[chore] 어드민 인증의 의미를 강조하기 위해 AdminAuthController로 이름 변경(#46)
기존 이름은 Admin 관련 모든 작업을 수행하는 듯한 느낌을 주지만, 실제로는 인증 관련 작업만 수행하므로 의미를 명확하게 하기 위해 이름 변경
